refactor(get-event): refat get event request
