+++ back/event/src/modules/get_event/get_event_controller.ts
@@ -1,40 +1,48 @@
-import { GetEventUsecase } from "./get_event_usecase"
+import { httpStatusCode } from '../../shared/domain/helpers/http'
+import { GetEventUsecase } from './get_event_usecase'
+import { GetEventRequest, GetEventResponse } from './protocols'
 
-export interface GetEventControllerProps {
-    usecase: GetEventUsecase
-    call: (req: any) => any
-  }
-  
-export class GetEventController implements GetEventControllerProps {
-    constructor(public usecase: GetEventUsecase) {
-      this.usecase = usecase
-    }
-  
-    call(req: any) {
-        const id = req.params.id
-        if (id === undefined) {
-            return {
-                "status": 400,
-                "message": "id is undefined",
-                "event": undefined 
-            }
-        }
-        const event = this.usecase.call(id)
-        if (event === undefined) {
-            return {
-                "status": 404,
-                "message": "event not found",
-                "event": undefined 
-            }
-        }
-        else {
-            return {
-                "status": 200,
-                "message": "event found",
-                "event": event.to_json()
-            }
-        }
-    }
+interface GetEventControllerProps {
+  usecase: GetEventUsecase
+  call(req: GetEventRequest): Promise<GetEventResponse>
 }
 
-  
+export class GetEventController implements GetEventControllerProps {
+  constructor(public usecase: GetEventUsecase) {
+    this.usecase = usecase
+  }
+
+  async call(req: GetEventRequest) {
+    const id = req.params.id
+
+    if (!id) {
+      return {
+        status: httpStatusCode.BAD_REQUEST,
+        message: 'id is undefined'
+      }
+    }
+
+    try {
+      const event = await this.usecase.call(id)
+
+      return {
+        status: httpStatusCode.OK,
+        message: 'event found',
+        event: event.toJson()
+      }
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } catch (error: any) {
+      if (error.message.includes('not found')) {
+        return {
+          status: httpStatusCode.NOT_FOUND,
+          message: error.message
+        }
+      }
+
+      return {
+        status: httpStatusCode.INTERNAL_SERVER_ERROR,
+        message: error.message
+      }
+    }
+  }
+}
