+++ back/event/src/modules/get_event/get_event_usecase.ts
@@ -1,19 +1,17 @@
-import { EventRepositoryInterface } from "../../shared/domain/repos/event_repository_interface"
-import { Event } from "../../shared/domain/entities/event"
+import { EventRepositoryInterface } from '../../shared/infra/repos/event_repository_interface'
+import { Event } from '../../shared/domain/entities/event'
 
 export interface GetEventUsecaseProps {
-    repo: EventRepositoryInterface
-    call: (id: string) => Event | undefined
-  }
-  
-export class GetEventUsecase implements GetEventUsecaseProps {
-    constructor(public repo: EventRepositoryInterface) {
-      this.repo = repo
-    }
-  
-    call(id: string) {
-        return this.repo.getEvent(id)
-    }
+  repo: EventRepositoryInterface
+  call(id: string): Promise<Event>
 }
 
-  
+export class GetEventUsecase implements GetEventUsecaseProps {
+  constructor(public repo: EventRepositoryInterface) {
+    this.repo = repo
+  }
+
+  async call(id: string) {
+    return await this.repo.getEvent(id)
+  }
+}
